+++ src/Calculator/index.js
@@ -1,4 +1,5 @@
 import React from 'react';
+import PriceQuantity from './PriceQuantity';
 
 export default class Calculator extends React.Component {
     state = {
@@ -35,7 +36,7 @@ export default class Calculator extends React.Component {
         const averagePriceKeys = Object.keys(this.state.averagePrice);
         const priceQuantityArr = averagePriceKeys.map((key) =>
             <React.Fragment key={key}>
-                <PriceQuantity index={key} handleChange={this.handleChange} data={this.state.averagePrice[key]} deleteRow={this.deleteRow}
+                <PriceQuantity index={key} handleChange={this.handleChange} deleteRow={this.deleteRow}
                 removeButton={averagePriceKeys.length > 1} />
                 <br />
             </React.Fragment>
@@ -57,49 +58,6 @@ export default class Calculator extends React.Component {
                 {priceQuantityArr.map(priceQuantity => priceQuantity)}
                 <button onClick={this.addRow}>Add Row</button>
             </div>
-        );
-    }
-}
-
-class PriceQuantity extends React.Component {
-    state = {
-        price: 0.0,
-        quantity: 0,
-        totalPrice: 0.0
-    };
-
-    updateTotalPrice = () => {
-        this.setState(state => ({totalPrice: state.price * state.quantity}));
-        this.setState(state => {
-            const totalPrice = state.price * state.quantity;
-            this.props.handleChange(this.props.index, totalPrice, state.quantity);
-            return {totalPrice};
-        })
-    }
-    
-    handlePriceChange = event => {
-        const value = event.target.value
-        this.setState({price: value || 0.0});
-        this.updateTotalPrice();
-    }
-   
-    handleQuantityChange = event => {
-        const value = parseInt(event.target.value, 0)
-        this.setState({quantity: value});
-        this.updateTotalPrice();
-    }
-
-    render() {
-        const removeButton = this.props.removeButton? <button key-id={this.props.index} onClick={this.props.deleteRow}>X</button> : '';
-        return(
-            <React.Fragment>
-                <input type="number" name="price" placeholder="Price"
-                onChange={this.handlePriceChange} />{' * '}
-                <input type="number" name="quantity" placeholder="Quantity"
-                onChange={this.handleQuantityChange} />
-                {` = ${this.state.totalPrice} `}
-                {removeButton}
-            </React.Fragment>
         );
     }
 }
